pyproject.toml
[project]
name = "practical-fastapi"
version = "0.1.0"
description = "ぼくのかんがえたさいきょうのFastAPI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13.3",
    "authlib>=1.3.2",
    "fastapi>=0.115.0",
    "freezegun>=1.5.1",
    "httpx>=0.27.2",
    "passlib>=1.7.4",
    "psycopg[binary,pool]>=3.2.3",
    "pydantic-settings>=2.5.2",
    "pyjwt>=2.9.0",
    "pytest-cov>=5.0.0",
    "python-multipart>=0.0.12",
    "sqlalchemy[postgresql-asyncpg]>=2.0.35",
    "uvicorn>=0.31.0",
]

[tool.uv]
dev-dependencies = [
    "factory-boy>=3.3.1",
    "pytest-asyncio>=0.24.0",
    "pytest>=8.3.3",
    "ruff>=0.6.9",
    "tqdm>=4.66.5",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__init__.py",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "migrations",
    "scripts",
    "settings.py",
    "tests",
]
line-length = 100

[tool.ruff.format]
quote-style = "double"

[tool.pytest]
args = """
  --tb short
"""
